# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/group/create": {
            "post": {
                "description": "根据请求参数创建一个新的群聊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "创建新群聊",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GroupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "成功创建群聊",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/response.GroupInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误或创建失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/group/delete": {
            "post": {
                "description": "根据请求参数删除群聊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "解散群聊",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功解散群聊",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误或解散失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/group/find": {
            "get": {
                "description": "通过群名模糊搜索所有的小组",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "模糊搜索所有的小组",
                "responses": {
                    "200": {
                        "description": "查询到群聊",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/response.GroupInfos"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "出现错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/group/info/{groupNum}": {
            "get": {
                "description": "根据群号获取群聊的详细信息，包括人数、群聊名称和群号",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "获取群聊信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "群号",
                        "name": "groupNUM",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取群聊信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.GroupInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误或获取失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/group/list/{id}": {
            "get": {
                "description": "通过用户id获取新的自己的小组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "通过用户id获取新的自己的小组",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取群成员列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.GroupInfos"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误或获取失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/group/member/add": {
            "post": {
                "description": "通过用户id和群号加入群聊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "加入群聊",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ExecuteGroupMember"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功加入群聊",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误或加入失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/group/member/check": {
            "post": {
                "description": "通过用户id和群号判断用户是否加入群聊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "判断用户是否加入小组",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GroupMember"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户已加入该群聊",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "用户未加入群聊",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/group/member/delete": {
            "post": {
                "description": "通过用户id和群号退出群聊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "退出群聊",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ExecuteGroupMember"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功退出群聊",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误或退出失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/group/members/{groupNum}": {
            "get": {
                "description": "根据群号获取群成员列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "获取群成员列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "群号",
                        "name": "groupNum",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取群成员列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.Users"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误或获取失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/group/ten": {
            "get": {
                "description": "通过页码获取小组前十条信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "分页获取小组十条信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pn",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取小组前十条信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.GroupInfos"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误或获取失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/group/update": {
            "post": {
                "description": "根据请求参数更新群聊的信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "更新群聊相关信息",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功更新群聊信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/response.GroupInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误或更新失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.ExecuteGroupMember": {
            "type": "object",
            "properties": {
                "group_num": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "request.GroupMember": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "request.GroupRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "execute_id": {
                    "description": "执行任务的用户id",
                    "type": "string"
                },
                "group_num": {
                    "description": "群号 = id",
                    "type": "string"
                },
                "is_public": {
                    "description": "任务是否公开",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "description": "群主设置",
                    "type": "string"
                }
            }
        },
...
        "response.GroupInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "group_owner": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_public": {
                    "description": "任务是否公开",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "response.GroupInfos": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.GroupInfo"
                    }
                },
                "nums": {
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "response.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "goal_public": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "slogan": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "response.Users": {
            "type": "object",
            "properties": {
                "nums": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.User"
                    }
                }
            }
        }
    }
}